app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic>=2.0
pytest
httpx
python-dotenv
python-multipart
requests
pyodbc

azure-core
azure-cognitiveservices-speech
pydantic-settings
